Flatten conversation module into a single file
