--- chatterbot/conversation/__init__.py
@@ -1,8 +1,0 @@
-from .statement import Statement
-from .response import Response
-
-
-__all__ = (
-    'Statement',
-    'Response'
-)
